feat(#35): created a seperate extractUser function, refactored code as per PR comments and refactored the test cases as per the latest changes
